// packages/client/src/store/profile/actionTypes.ts
export enum ProfileActionTypes {
  GetBasic = 'PROFILE.GET_BASIC',
  GetIntermediate = 'PROFILE.GET_INTERMEDIATE',
  GetAdvanced = 'PROFILE.GET_ADVANCED',
  SubmitBasic = 'PROFILE.SUBMIT_BASIC',
  SetBasic = 'PROFILE.SET_BASIC',
  UploadIntermediate = 'PROFILE.UPLOAD_INTERMEDIATE',
  UploadAdvanced = 'PROFILE.UPLOAD_ADVANCED',
  SetIntermediate = 'PROFILE.SET_INTERMEDIATE',
  SetAdvanced = 'PROFILE.SET_ADVANCED',
  Withdraw = 'PROFILE.WITHDRAW',
  WithdrawConfirm = 'PROFILE.WITHDRAW_CONFIRM',
  WithdrawRequestCode = 'PROFILE.REQUEST_CODE',
  DropKYCFile = 'PROFILE.DROP_KYC_FILE',
  GetTransactionAll = 'PROFILE.GET_TRANSACTION_ALL',
  SetTransactionAll = 'PROFILE.SET_TRANSACTION_ALL',
  SetWithdraw = 'PROFILE.SET_WITHDRAW',
  GetSportsBet = 'PROFILE.GET_SPORTS_BET',
  SetSportsBet = 'PROFILE.SET_SPORTS_BET',
}